With this patch, libssh 0.5.5 now builds with gcc 4.1.2.
Silly stuff.

They are printing out the usage for a sample program and are using a
concatenated string to do it. They have a couple of #ifdef's which control
whether or not a line is printed in the usage, and this is what gcc 4.1.2
is objecting to.

I just removed the #ifdef's, since 1) it's a sample program that we don't
use, and 2) all that happens is that the usage may print out some options
which aren't actually available, depending on the build.



git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@928 a9e16012-fd73-49ee-95cb-036de7809213