Beginning deprecation process for -2 option to jenkinsbild and mkfooall.sh/bildopts.sh.  The -2 option still works, but prints a WARNING message.
Added -i option to jenkinsbild, to specify that test results should
be ignored when installing. This option passes -I to mkfooall-default.sh

STEP 2 to jenkinsbold (either -2 or -s 2) no longer adds a -2 to the 
options sent to mkfooall-default.sh, so -F and -I are no longer added
to the options of mkfooall.sh.

This means that JRC should add the -i option to the second invocation
of jenkinsbild in his jenkins jobs.



git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@1257 a9e16012-fd73-49ee-95cb-036de7809213