Changed hard-coded "Program Files" to $programfiles so this works with Express version of Visual Studio.
Collapses setVs9Vars/setVs10Vars/setVs11Vars/setVs12Vars down to setVsVars()
with an argument for the version.

Also collapsed check for VISUALSTUDIO_VERSION if it's not already set.
 
Tested on VS12 and VS9 and it seems to work.

One issue is that $MINGW_BINDIR ends up in PATH twice, so need to sort that out.


git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@2095 a9e16012-fd73-49ee-95cb-036de7809213