Moving logic from machines/linux.intel into bildfcns.sh.
Intel is moving to allow mkl with non-intel compilers.  This allows
planning for that.  Note that this approach is to specify specific mkl
libraries.  For vorpal, this seems to be problematic so we have
CMakeLists.txt to switch to the -mkl=cluster format for compiling
linking.  This is not extensible to the gcc case but is sufficient for
now.


git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@2594 a9e16012-fd73-49ee-95cb-036de7809213